feat: source property addition to existing services

+++ app/routers/data_reference.py
@@ -11,8 +11,9 @@ from app.models.data_domain import DataDomain
 from app.models.fhir.bundle import Bundle
 from app.models.fhir.resources.data import (
     CARE_CONTEXT_SYSTEM,
+    ORG_SYSTEM,
+    ORG_TYPE_SYSTEM,
     SOURCE_SYSTEM,
-    SOURCE_TYPE_SYSTEM,
     SUBJECT_SYSTEM,
 )
 from app.models.fhir.resources.data_reference.requests import DataReferenceRequestParams
@@ -72,18 +73,22 @@ def exchange_oprf(pseudonym_service: PseudonymService, oprf_jwe: str, blind_fact
                                     "versionId": "1",
                                     "lastUpdated": "2024-12-08T14:40:00Z",
                                 },
-                                "source": {
-                                    "system": SOURCE_SYSTEM,
+                                "organization": {
+                                    "system": ORG_SYSTEM,
                                     "value": "90000001",
                                 },
-                                "sourceType": {
+                                "organization_type": {
                                     "coding": [
                                         {
-                                            "system": SOURCE_TYPE_SYSTEM,
+                                            "system": ORG_TYPE_SYSTEM,
                                             "code": "laboratorium",
                                             "display": "Laboratorium",
                                         }
                                     ]
+                                },
+                                "source": {
+                                    "system": SOURCE_SYSTEM,
+                                    "value": "Electronic Health Record System",
                                 },
                                 "careContext": {
                                     "coding": [
@@ -136,6 +141,7 @@ def get_reference(
     registrations = referral_service.get_registrations(
         ura_number=UraNumber(params.source),
         pseudonym=pseudonym,
+        source=params.source,
         data_domain=DataDomain(params.care_context) if params.care_context else None,
     )
     return Bundle.from_reference_outputs(registrations)
@@ -187,14 +193,16 @@ def delete_reference(
                 ura_number=UraNumber(params.source),
                 pseudonym=pseudonym,
             )
-        else:
+            return DeleteResponse(status_code=204)
+        if params.source and params.care_context:
             referral_service.delete_specific_registration(
-                ura_number=UraNumber(params.source),
+                ura_number=UraNumber(params.organization),
                 data_domain=DataDomain(params.care_context),
                 pseudonym=pseudonym,
+                source=params.source,
             )
 
-        return DeleteResponse(status_code=204)
+            return DeleteResponse(status_code=204)
 
     referral_service.delete_specific_organization(ura_number=UraNumber(params.source))
     return DeleteResponse(status_code=204)
@@ -221,19 +229,20 @@ def delete_reference(
                                 "versionId": "1",
                                 "lastUpdated": "2024-12-08T14:40:00Z",
                             },
-                            "source": {
-                                "system": SOURCE_SYSTEM,
+                            "organization": {
+                                "system": ORG_TYPE_SYSTEM,
                                 "value": "90000001",
                             },
-                            "sourceType": {
+                            "organization_typeType": {
                                 "coding": [
                                     {
-                                        "system": SOURCE_TYPE_SYSTEM,
+                                        "system": ORG_TYPE_SYSTEM,
                                         "code": "laboratorium",
                                         "display": "Laboratorium",
                                     }
                                 ]
                             },
+                            "source": {"system": SOURCE_SYSTEM, "value": "DeviceName"},
                             "careContext": {
                                 "coding": [
                                     {
@@ -267,19 +276,20 @@ def delete_reference(
                                 "versionId": "1",
                                 "lastUpdated": "2024-12-08T14:40:00Z",
                             },
-                            "source": {
-                                "system": SOURCE_SYSTEM,
+                            "organization": {
+                                "system": ORG_SYSTEM,
                                 "value": "90000001",
                             },
-                            "sourceType": {
+                            "organizationType": {
                                 "coding": [
                                     {
-                                        "system": SOURCE_TYPE_SYSTEM,
+                                        "system": ORG_TYPE_SYSTEM,
                                         "code": "laboratorium",
                                         "display": "Laboratorium",
                                     }
                                 ]
                             },
+                            "source": {"system": SOURCE_SYSTEM, "value": "DeviceValue"},
                             "careContext": {
                                 "coding": [
                                     {
@@ -313,15 +323,16 @@ def create_reference(
                         "system": SUBJECT_SYSTEM,
                         "value": "eyJhbGciOiJSU0EtT0FFUC0yNTYiLCJlbmMiOiJBMjU2R0NNIn0...",
                     },
-                    "source": {"system": SOURCE_SYSTEM, "value": "90000001"},
-                    "sourceType": {
+                    "organization": {"system": ORG_SYSTEM, "value": "90000001"},
+                    "organizationType": {
                         "coding": [
                             {
-                                "system": SOURCE_TYPE_SYSTEM,
+                                "system": ORG_TYPE_SYSTEM,
                                 "code": "laboratorium",
                             }
                         ]
                     },
+                    "source": {"system": SOURCE_SYSTEM, "value": "DeviceValue"},
                     "careContext": {
                         "coding": [
                             {
@@ -362,6 +373,7 @@ def create_reference(
         pseudonym=pseudonym,
         data_domain=data.get_data_domain(),
         ura_number=data.get_ura_number(),
+        source=data.get_source(),
     )
     if referral:
         return FHIRJSONResponse(
@@ -374,6 +386,7 @@ def create_reference(
         data_domain=data.get_data_domain(),
         ura_number=data.get_ura_number(),
         organization_type=data.get_organization_type(),
+        source=data.get_source(),
     )
     return FHIRJSONResponse(
         content=jsonable_encoder(new_reference.model_dump(exclude_none=True, by_alias=True)),
@@ -409,7 +422,7 @@ def create_reference(
                         "sourceType": {
                             "coding": [
                                 {
-                                    "system": SOURCE_TYPE_SYSTEM,
+                                    "system": ORG_TYPE_SYSTEM,
                                     "code": "laboratorium",
                                     "display": "Laboratorium",
                                 }
